+++ src/vs/editor/contrib/inlineCompletions/browser/model/inlineCompletionsSource.ts
@@ -16,7 +16,7 @@ import { IContextKeyService } from '../../../../../platform/contextkey/common/co
 import { IInstantiationService } from '../../../../../platform/instantiation/common/instantiation.js';
 import { ILogService } from '../../../../../platform/log/common/log.js';
 import { observableConfigValue } from '../../../../../platform/observable/common/platformObservableUtils.js';
-import { OffsetEdit, SingleOffsetEdit } from '../../../../common/core/offsetEdit.js';
+import { applyEditsToRanges, OffsetEdit, SingleOffsetEdit } from '../../../../common/core/offsetEdit.js';
 import { OffsetRange } from '../../../../common/core/offsetRange.js';
 import { Position } from '../../../../common/core/position.js';
 import { Range } from '../../../../common/core/range.js';
@@ -761,71 +761,4 @@ export class StructuredLogger<T extends IRecordableLogEntry> extends Disposable 
 
 export function observableContextKey<T>(key: string, contextKeyService: IContextKeyService): IObservable<T | undefined> {
 	return observableFromEvent(contextKeyService.onDidChangeContext, () => contextKeyService.getContextKeyValue<T>(key));
-}
-
-export function applyEditsToRanges(sortedRanges: OffsetRange[], edits: OffsetEdit): OffsetRange[] {
-	sortedRanges = sortedRanges.slice();
-
-	// treat edits as deletion of the replace range and then as insertion that extends the first range
-	const result: OffsetRange[] = [];
-
-	let offset = 0;
-
-	for (const e of edits.edits) {
-		while (true) {
-			// ranges before the current edit
-			const r = sortedRanges[0];
-			if (!r || r.endExclusive >= e.replaceRange.start) {
-				break;
-			}
-			sortedRanges.shift();
-			result.push(r.delta(offset));
-		}
-
-		const intersecting: OffsetRange[] = [];
-		while (true) {
-			const r = sortedRanges[0];
-			if (!r || !r.intersectsOrTouches(e.replaceRange)) {
-				break;
-			}
-			sortedRanges.shift();
-			intersecting.push(r);
-		}
-
-		for (let i = intersecting.length - 1; i >= 0; i--) {
-			let r = intersecting[i];
-
-			const overlap = r.intersect(e.replaceRange)!.length;
-			r = r.deltaEnd(-overlap + (i === 0 ? e.newText.length : 0));
-
-			const rangeAheadOfReplaceRange = r.start - e.replaceRange.start;
-			if (rangeAheadOfReplaceRange > 0) {
-				r = r.delta(-rangeAheadOfReplaceRange);
-			}
-
-			if (i !== 0) {
-				r = r.delta(e.newText.length);
-			}
-
-			// We already took our offset into account.
-			// Because we add r back to the queue (which then adds offset again),
-			// we have to remove it here.
-			r = r.delta(-(e.newText.length - e.replaceRange.length));
-
-			sortedRanges.unshift(r);
-		}
-
-		offset += e.newText.length - e.replaceRange.length;
-	}
-
-	while (true) {
-		const r = sortedRanges[0];
-		if (!r) {
-			break;
-		}
-		sortedRanges.shift();
-		result.push(r.delta(offset));
-	}
-
-	return result;
 }
